[mapillary] Clean up build.gradle
The build file for Gradle is now split in several parts for a better overview and better maintainability.
The files `gradle/*.gradle` are parts that are unlikely to change often, because these are not specific to this plugin and can be configured through `gradle.properties`.
Also the Gradle version is bumped to 2.14.1, because a critical defect has been fixed since 2.14 (see https://docs.gradle.org/2.14.1/release-notes).

git-svn-id: http://svn.openstreetmap.org/applications/editors/josm/plugins@32693 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4